+++ BankProject/BankProject/ConvertToDollar.cs
@@ -9,14 +9,14 @@ namespace BankProject
 {
     internal class ConvertToDollar : CurrencyConverter
     {
-        protected float ConvertionRateDollar = 10.51f;
-        public ConvertToDollar(float currencyAmount) : base(currencyAmount)
+        protected double ConvertionRateDollar = 10.51f;
+        public ConvertToDollar(double currencyAmount) : base(currencyAmount)
         {
             this.ConvertionRate = ConvertionRateDollar;
         }
 
 
-        public override string PrintCurrencyConverter(float currencyAmount)
+        public override string PrintCurrencyConverter(double currencyAmount)
         {
             return $"\n\tExchange Rate: {ConvertionRateDollar}\n\tExchange Value USS:{CurrencyConverterCalculator(currencyAmount).ToString("c2", CultureInfo.CreateSpecificCulture("en-us"))}";
         }
