+++ BankProject/BankProject/ConvertToEuro.cs
@@ -9,14 +9,14 @@ namespace BankProject
 {
     internal class ConvertToEuro : CurrencyConverter
     {
-        protected float ConvertionRateEuro = 11.17f;
-        public ConvertToEuro(float currencyAmount) : base(currencyAmount)
+        protected double ConvertionRateEuro = 11.17f;
+        public ConvertToEuro(double currencyAmount) : base(currencyAmount)
         {
             this.ConvertionRate = ConvertionRateEuro;
         }
 
 
-        public override string PrintCurrencyConverter(float currencyAmount)
+        public override string PrintCurrencyConverter(double currencyAmount)
         {
             return $"\n\tExchange Rate: {ConvertionRateEuro}\n\tExchange Value EUR:{CurrencyConverterCalculator(currencyAmount).ToString("c2", CultureInfo.CreateSpecificCulture("fr-FR"))}";
         }
